polish(tests): hoist probe_env_to_paths duplicates into test_support helpers
`embed/tests.rs::t_018_*` and `reranker/tests.rs::t_019_*` were ~50-line
copy-pastes that differed only in the `<EmbedKind>` / `<RerankerKind>`
type argument; the same applied to the
`probe_env_to_paths_returns_*_when_all_present` pair.

Lifted both shapes into `K: ModelKind` generic helpers under
`crate::test_support`:

- `assert_probe_env_to_paths_preserves_snapshot_symlink_filename<K>`
  (Issue #107 regression — load path must remain the snapshot symlink,
  not the canonicalized blob path)
- `assert_probe_env_to_paths_returns_paths_when_all_present<K>`
  (paths field round-trips uncanonicalized)

The kind-specific tests now reduce to one-line wrappers carrying only
the type argument, so any future invariant change lands in a single
place. `cargo test --lib`: 317 passed, 3 ignored.
